rename api-gateway/consul package to api-gateway/translation
