+++ control-plane/api-gateway/translation/config_entry_translation_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package consul
+package translation
 
 import (
 	"testing"
